requirements.txt
pandas
scikit-learn
matplotlib
pyserial
